fix(securityhub): tests
