+++ prowler/lib/outputs/json_asff/json_asff.py
@@ -2,6 +2,7 @@ from prowler.config.config import prowler_version, timestamp_utc
 from prowler.lib.logger import logger
 from prowler.lib.outputs.compliance.compliance import get_check_compliance
 from prowler.lib.outputs.json_asff.models import (
+    Check_Output_JSON_ASFF,
     Compliance,
     ProductFields,
     Resource,
@@ -44,51 +45,34 @@ def generate_json_asff_resource_tags(tags):
         )
 
 
-def fill_json_asff(finding_output, provider, finding, output_options):
+def fill_json_asff(provider, finding):
+    """
+    Fill the finding's output in JSON ASFF format.
+
+    Parameters:
+    - provider: The provider object containing information about the provider (e.g., AWS) and the output options object containing information about the desired output format.
+    - finding: The finding object containing information about the specific finding.
+
+    Returns:
+    - finding_output: The filled finding's output in JSON ASFF format.
+    """
+
     try:
         # Check if there are no resources in the finding
         if finding.resource_arn == "":
             if finding.resource_id == "":
                 finding.resource_id = "NONE_PROVIDED"
             finding.resource_arn = finding.resource_id
-        # The following line cannot be changed because it is the format we use to generate unique findings for AWS Security Hub
-        # If changed some findings could be lost because the unique identifier will be different
-        # TODO: get this from the provider output
-        finding_output.Id = f"prowler-{finding.check_metadata.CheckID}-{provider.identity.account}-{finding.region}-{hash_sha512(finding.resource_id)}"
-        finding_output.ProductArn = f"arn:{provider.identity.partition}:securityhub:{finding.region}::product/prowler/prowler"
-        finding_output.ProductFields = ProductFields(
-            ProviderVersion=prowler_version, ProwlerResourceName=finding.resource_arn
-        )
-        finding_output.GeneratorId = "prowler-" + finding.check_metadata.CheckID
-        finding_output.AwsAccountId = provider.identity.account
-        finding_output.Types = finding.check_metadata.CheckType
-        finding_output.FirstObservedAt = finding_output.UpdatedAt = (
-            finding_output.CreatedAt
-        ) = timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
-        finding_output.Severity = Severity(
-            Label=finding.check_metadata.Severity.upper()
-        )
-        finding_output.Title = finding.check_metadata.CheckTitle
-        # Description should NOT be longer than 1024 characters
-        finding_output.Description = (
-            (finding.status_extended[:1000] + "...")
-            if len(finding.status_extended) > 1000
-            else finding.status_extended
-        )
+
+        timestamp = timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
         resource_tags = generate_json_asff_resource_tags(finding.resource_tags)
-        finding_output.Resources = [
-            Resource(
-                Id=finding.resource_arn,
-                Type=finding.check_metadata.ResourceType,
-                Partition=provider.identity.partition,
-                Region=finding.region,
-                Tags=resource_tags,
-            )
-        ]
+
         # Iterate for each compliance framework
         compliance_summary = []
         associated_standards = []
-        check_compliance = get_check_compliance(finding, "aws", output_options)
+        check_compliance = get_check_compliance(
+            finding, provider.type, provider.output_options
+        )
         for key, value in check_compliance.items():
             if (
                 len(associated_standards) < 20
@@ -102,19 +86,49 @@ def fill_json_asff(finding_output, provider, finding, output_options):
         # Ensures finding_status matches allowed values in ASFF
         finding_status = generate_json_asff_status(finding.status)
 
-        finding_output.Compliance = Compliance(
-            Status=finding_status,
-            AssociatedStandards=associated_standards,
-            RelatedRequirements=compliance_summary,
+        json_asff_output = Check_Output_JSON_ASFF(
+            # The following line cannot be changed because it is the format we use to generate unique findings for AWS Security Hub
+            # If changed some findings could be lost because the unique identifier will be different
+            # TODO: get this from the provider output
+            Id=f"prowler-{finding.check_metadata.CheckID}-{provider.identity.account}-{finding.region}-{hash_sha512(finding.resource_id)}",
+            ProductArn=f"arn:{provider.identity.partition}:securityhub:{finding.region}::product/prowler/prowler",
+            ProductFields=ProductFields(
+                ProviderVersion=prowler_version,
+                ProwlerResourceName=finding.resource_arn,
+            ),
+            GeneratorId="prowler-" + finding.check_metadata.CheckID,
+            AwsAccountId=provider.identity.account,
+            Types=finding.check_metadata.CheckType,
+            FirstObservedAt=timestamp,
+            UpdatedAt=timestamp,
+            CreatedAt=timestamp,
+            Severity=Severity(Label=finding.check_metadata.Severity.upper()),
+            Title=finding.check_metadata.CheckTitle,
+            # Description should NOT be longer than 1024 characters
+            Description=(
+                (finding.status_extended[:1000] + "...")
+                if len(finding.status_extended) > 1000
+                else finding.status_extended
+            ),
+            Resources=[
+                Resource(
+                    Id=finding.resource_arn,
+                    Type=finding.check_metadata.ResourceType,
+                    Partition=provider.identity.partition,
+                    Region=finding.region,
+                    Tags=resource_tags,
+                )
+            ],
+            Compliance=Compliance(
+                Status=finding_status,
+                AssociatedStandards=associated_standards,
+                RelatedRequirements=compliance_summary,
+            ),
+            Remediation={
+                "Recommendation": finding.check_metadata.Remediation.Recommendation
+            },
         )
-        # Fill Recommendation Url if it is blank
-        if not finding.check_metadata.Remediation.Recommendation.Url:
-            finding.check_metadata.Remediation.Recommendation.Url = "https://docs.aws.amazon.com/securityhub/latest/userguide/what-is-securityhub.html"
-        finding_output.Remediation = {
-            "Recommendation": finding.check_metadata.Remediation.Recommendation
-        }
-
-        return finding_output
+        return json_asff_output
     except Exception as error:
         logger.error(
             f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
