+++ tests/lib/outputs/outputs_test.py
@@ -1,11 +1,9 @@
 import os
 from os import path, remove
-from time import mktime
 from unittest import mock
 
 import pytest
 from colorama import Fore
-from mock import patch
 
 from prowler.config.config import (
     csv_file_suffix,
@@ -13,9 +11,6 @@ from prowler.config.config import (
     json_file_suffix,
     orange_color,
     output_file_timestamp,
-    prowler_version,
-    timestamp,
-    timestamp_utc,
 )
 from prowler.lib.check.compliance_models import (
     CIS_Requirement_Attribute,
@@ -23,76 +18,27 @@ from prowler.lib.check.compliance_models import (
     Compliance_Requirement,
 )
 from prowler.lib.check.models import Check_Report, load_check_metadata
+from prowler.lib.outputs.common_models import FindingOutput
+from prowler.lib.outputs.compliance.compliance import get_check_compliance
+from prowler.lib.outputs.csv.csv import generate_csv_fields
 from prowler.lib.outputs.file_descriptors import fill_file_descriptors
-from prowler.lib.outputs.json import (
-    fill_json_asff,
-    fill_json_ocsf,
-    generate_json_asff_resource_tags,
-    generate_json_asff_status,
-)
-from prowler.lib.outputs.models import (
-    Account,
-    Check_Output_CSV,
-    Check_Output_JSON_ASFF,
-    Check_Output_JSON_OCSF,
-    Cloud,
-    Compliance,
-    Compliance_OCSF,
-    Feature,
-    Finding,
-    Group,
-    Metadata,
-    Organization,
-    Product,
-    ProductFields,
-    Remediation_OCSF,
-    Resource,
-    Resources,
-    Severity,
-    generate_csv_fields,
-    get_check_compliance,
+from prowler.lib.outputs.outputs import extract_findings_statistics, set_report_color
+from prowler.lib.outputs.utils import (
     parse_json_tags,
     unroll_dict,
     unroll_dict_to_list,
     unroll_list,
     unroll_tags,
 )
-from prowler.lib.outputs.outputs import extract_findings_statistics, set_report_color
-from prowler.lib.utils.utils import hash_sha512, open_file
-from prowler.providers.common.models import Audit_Metadata
-
-AWS_ACCOUNT_ID = "123456789012"
+from prowler.lib.utils.utils import open_file
+from tests.providers.aws.utils import AWS_ACCOUNT_NUMBER, set_mocked_aws_provider
 
 
-class Test_Outputs:
-    def test_fill_file_descriptors(self):
-        audited_account = AWS_ACCOUNT_ID
+class TestOutputs:
+    def test_fill_file_descriptors_aws(self):
+        audited_account = AWS_ACCOUNT_NUMBER
         output_directory = f"{os.path.dirname(os.path.realpath(__file__))}"
-        # TODO(Audit_Info): use provider here
-        audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
+        aws_provider = set_mocked_aws_provider()
         test_output_modes = [
             ["csv"],
             ["json-asff"],
@@ -140,7 +86,7 @@ class Test_Outputs:
                 output_mode_list,
                 output_directory,
                 output_filename,
-                audit_info,
+                aws_provider,
             )
             for output_mode in output_mode_list:
                 assert (
@@ -199,7 +145,7 @@ class Test_Outputs:
             "notes",
         ]
 
-        assert generate_csv_fields(Check_Output_CSV) == expected
+        assert generate_csv_fields(FindingOutput) == expected
 
     def test_unroll_list(self):
         list = ["test", "test1", "test2"]
@@ -305,9 +251,116 @@ class Test_Outputs:
         assert parse_json_tags([{}]) == {}
         assert parse_json_tags(None) == {}
 
-    # def test_fill_json(self):
+
+class TestOutputJSONOCSF:
+    # def test_fill_json_ocsf_iso_format_timestamp(self):
+    #     aws_provider = set_mocked_aws_provider()
+    #     finding = Check_Report(
+    #         load_check_metadata(
+    #             f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
+    #         ).json()
+    #     )
+    #     finding.resource_details = "Test resource details"
+    #     finding.resource_id = "test-resource"
+    #     finding.resource_arn = "test-arn"
+    #     finding.region = "eu-west-1"
+    #     finding.status = "PASS"
+    #     finding.status_extended = "This is a test"
+    #     # TODO: fill_json_ocsf
+    #     expected = Check_Output_JSON_OCSF(
+    #         finding=Finding(
+    #             title="Ensure Access Keys unused are disabled",
+    #             desc="Ensure Access Keys unused are disabled",
+    #             supporting_data={
+    #                 "Risk": "Risk associated.",
+    #                 "Notes": "additional information",
+    #             },
+    #             remediation=Remediation_OCSF(
+    #                 kb_articles=[
+    #                     "code or URL to the code location.",
+    #                     "code or URL to the code location.",
+    #                     "cli command or URL to the cli command location.",
+    #                     "cli command or URL to the cli command location.",
+    #                     "https://myfp.com/recommendations/dangerous_things_and_how_to_fix_them.html",
+    #                 ],
+    #                 desc="Run sudo yum update and cross your fingers and toes.",
+    #             ),
+    #             types=["Software and Configuration Checks"],
+    #             src_url="https://serviceofficialsiteorpageforthissubject",
+    #             uid="prowler-aws-iam_user_accesskey_unused-123456789012-eu-west-1-test-resource",
+    #             related_events=[
+    #                 "othercheck1",
+    #                 "othercheck2",
+    #                 "othercheck3",
+    #                 "othercheck4",
+    #             ],
+    #         ),
+    #         resources=[
+    #             Resources(
+    #                 group=Group(name="iam"),
+    #                 region="eu-west-1",
+    #                 name="test-resource",
+    #                 uid="test-arn",
+    #                 labels=[],
+    #                 type="AwsIamAccessAnalyzer",
+    #                 details="Test resource details",
+    #             )
+    #         ],
+    #         status_detail="This is a test",
+    #         compliance=Compliance_OCSF(
+    #             status="Success", requirements=[], status_detail="This is a test"
+    #         ),
+    #         message="This is a test",
+    #         severity_id=2,
+    #         severity="Low",
+    #         cloud=Cloud(
+    #             account=Account(name="", uid="123456789012"),
+    #             region="eu-west-1",
+    #             org=Organization(uid="", name=""),
+    #             provider="aws",
+    #             project_uid="",
+    #         ),
+    #         time=timestamp.isoformat(),
+    #         metadata=Metadata(
+    #             original_time=timestamp.isoformat(),
+    #             profiles=["default"],
+    #             product=Product(
+    #                 language="en",
+    #                 name="Prowler",
+    #                 version=prowler_version,
+    #                 vendor_name="Prowler/ProwlerPro",
+    #                 feature=Feature(
+    #                     name="iam_user_accesskey_unused",
+    #                     uid="iam_user_accesskey_unused",
+    #                     version=prowler_version,
+    #                 ),
+    #             ),
+    #             version="1.0.0-rc.3",
+    #         ),
+    #         state_id=0,
+    #         state="New",
+    #         status_id=1,
+    #         status="Success",
+    #         type_uid=200101,
+    #         type_name="Security Finding: Create",
+    #         impact_id=0,
+    #         impact="Unknown",
+    #         confidence_id=0,
+    #         confidence="Unknown",
+    #         activity_id=1,
+    #         activity_name="Create",
+    #         category_uid=2,
+    #         category_name="Findings",
+    #         class_uid=2001,
+    #         class_name="Security Finding",
+    #     )
+    #     output_options = mock.MagicMock()
+    #     output_options.unix_timestamp = False
+    #     assert fill_json_ocsf(aws_provider, finding, output_options) == expected
+
+    # def test_fill_json_ocsf_unix_timestamp(self):
     #     input_audit_info = AWS_Audit_Info(
-    #         session_config = None,
+    #         session_config=None,
     #         original_session=None,
     #         audit_session=None,
     #         audited_account=AWS_ACCOUNT_ID,
@@ -321,6 +374,14 @@ class Test_Outputs:
     #         assumed_role_info=None,
     #         audited_regions=["eu-west-2", "eu-west-1"],
     #         organizations_metadata=None,
+    #         audit_resources=None,
+    #         mfa_enabled=False,
+    #         audit_metadata=Audit_Metadata(
+    #             services_scanned=0,
+    #             expected_checks=[],
+    #             completed_checks=0,
+    #             audit_progress=0,
+    #         ),
     #     )
     #     finding = Check_Report(
     #         load_check_metadata(
@@ -334,819 +395,96 @@ class Test_Outputs:
     #     finding.status = "PASS"
     #     finding.status_extended = "This is a test"
 
-    #     input = Check_Output_JSON(**finding.check_metadata.dict())
-
-    #     expected = Check_Output_JSON(**finding.check_metadata.dict())
-    #     expected.AssessmentStartTime = timestamp_iso
-    #     expected.FindingUniqueId = ""
-    #     expected.Profile = "default"
-    #     expected.AccountId = AWS_ACCOUNT_ID
-    #     expected.OrganizationsInfo = None
-    #     expected.Region = "eu-west-1"
-    #     expected.Status = "PASS"
-    #     expected.StatusExtended = "This is a test"
-    #     expected.ResourceId = "test-resource"
-    #     expected.ResourceArn = "test-arn"
-    #     expected.ResourceDetails = "Test resource details"
-
-    #     assert fill_json(input, input_audit_info, finding) == expected
-
-    def test_fill_json_asff(self):
-        input_audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
-        finding = Check_Report(
-            load_check_metadata(
-                f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
-            ).json()
-        )
-        finding.resource_details = "Test resource details"
-        finding.resource_id = "test-resource"
-        finding.resource_arn = "test-arn"
-        finding.region = "eu-west-1"
-        finding.status = "PASS"
-        finding.status_extended = "This is a test"
-
-        expected = Check_Output_JSON_ASFF()
-        expected.Id = f"prowler-{finding.check_metadata.CheckID}-123456789012-eu-west-1-{hash_sha512('test-resource')}"
-        expected.ProductArn = "arn:aws:securityhub:eu-west-1::product/prowler/prowler"
-        expected.ProductFields = ProductFields(
-            ProviderVersion=prowler_version, ProwlerResourceName="test-arn"
-        )
-        expected.GeneratorId = "prowler-" + finding.check_metadata.CheckID
-        expected.AwsAccountId = AWS_ACCOUNT_ID
-        expected.Types = finding.check_metadata.CheckType
-        expected.FirstObservedAt = expected.UpdatedAt = expected.CreatedAt = (
-            timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
-        )
-        expected.Severity = Severity(Label=finding.check_metadata.Severity.upper())
-        expected.Title = finding.check_metadata.CheckTitle
-        expected.Description = finding.status_extended
-        expected.Resources = [
-            Resource(
-                Id="test-arn",
-                Type=finding.check_metadata.ResourceType,
-                Partition="aws",
-                Region="eu-west-1",
-            )
-        ]
-
-        expected.Compliance = Compliance(
-            Status="PASS" + "ED",
-            RelatedRequirements=[],
-            AssociatedStandards=[],
-        )
-        expected.Remediation = {
-            "Recommendation": finding.check_metadata.Remediation.Recommendation
-        }
-
-        input = Check_Output_JSON_ASFF()
-        output_options = mock.MagicMock()
-
-        assert (
-            fill_json_asff(input, input_audit_info, finding, output_options) == expected
-        )
-
-    def test_fill_json_asff_without_remediation_recommendation_url(self):
-        input_audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
-        finding = Check_Report(
-            load_check_metadata(
-                f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
-            ).json()
-        )
-
-        # Empty the Remediation.Recomendation.URL
-        finding.check_metadata.Remediation.Recommendation.Url = ""
-
-        finding.resource_details = "Test resource details"
-        finding.resource_id = "test-resource"
-        finding.resource_arn = "test-arn"
-        finding.region = "eu-west-1"
-        finding.status = "PASS"
-        finding.status_extended = "This is a test"
-
-        expected = Check_Output_JSON_ASFF()
-        expected.Id = f"prowler-{finding.check_metadata.CheckID}-123456789012-eu-west-1-{hash_sha512('test-resource')}"
-        expected.ProductArn = "arn:aws:securityhub:eu-west-1::product/prowler/prowler"
-        expected.ProductFields = ProductFields(
-            ProviderVersion=prowler_version, ProwlerResourceName="test-arn"
-        )
-        expected.GeneratorId = "prowler-" + finding.check_metadata.CheckID
-        expected.AwsAccountId = AWS_ACCOUNT_ID
-        expected.Types = finding.check_metadata.CheckType
-        expected.FirstObservedAt = expected.UpdatedAt = expected.CreatedAt = (
-            timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
-        )
-        expected.Severity = Severity(Label=finding.check_metadata.Severity.upper())
-        expected.Title = finding.check_metadata.CheckTitle
-        expected.Description = finding.status_extended
-        expected.Resources = [
-            Resource(
-                Id="test-arn",
-                Type=finding.check_metadata.ResourceType,
-                Partition="aws",
-                Region="eu-west-1",
-            )
-        ]
-
-        expected.Compliance = Compliance(
-            Status="PASS" + "ED",
-            RelatedRequirements=[],
-            AssociatedStandards=[],
-        )
-
-        # Set the check's remediation
-        expected.Remediation = {
-            "Recommendation": finding.check_metadata.Remediation.Recommendation,
-            # "Code": finding.check_metadata.Remediation.Code,
-        }
-
-        expected.Remediation["Recommendation"].Text = (
-            finding.check_metadata.Remediation.Recommendation.Text
-        )
-        expected.Remediation["Recommendation"].Url = (
-            "https://docs.aws.amazon.com/securityhub/latest/userguide/what-is-securityhub.html"
-        )
-
-        input = Check_Output_JSON_ASFF()
-        output_options = mock.MagicMock()
-
-        assert (
-            fill_json_asff(input, input_audit_info, finding, output_options) == expected
-        )
-
-    def test_fill_json_asff_with_long_description(self):
-        input_audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
-        finding = Check_Report(
-            load_check_metadata(
-                f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
-            ).json()
-        )
-
-        # Empty the Remediation.Recomendation.URL
-        finding.check_metadata.Remediation.Recommendation.Url = ""
-
-        finding.resource_details = "Test resource details"
-        finding.resource_id = "test-resource"
-        finding.resource_arn = "test-arn"
-        finding.region = "eu-west-1"
-        finding.status = "PASS"
-        finding.status_extended = "x" * 2000  # it has to be limited to 1000+...
-
-        expected = Check_Output_JSON_ASFF()
-        expected.Id = f"prowler-{finding.check_metadata.CheckID}-123456789012-eu-west-1-{hash_sha512('test-resource')}"
-        expected.ProductArn = "arn:aws:securityhub:eu-west-1::product/prowler/prowler"
-        expected.ProductFields = ProductFields(
-            ProviderVersion=prowler_version, ProwlerResourceName="test-arn"
-        )
-        expected.GeneratorId = "prowler-" + finding.check_metadata.CheckID
-        expected.AwsAccountId = AWS_ACCOUNT_ID
-        expected.Types = finding.check_metadata.CheckType
-        expected.FirstObservedAt = expected.UpdatedAt = expected.CreatedAt = (
-            timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
-        )
-        expected.Severity = Severity(Label=finding.check_metadata.Severity.upper())
-        expected.Title = finding.check_metadata.CheckTitle
-        expected.Description = finding.status_extended[:1000] + "..."
-        expected.Resources = [
-            Resource(
-                Id="test-arn",
-                Type=finding.check_metadata.ResourceType,
-                Partition="aws",
-                Region="eu-west-1",
-            )
-        ]
-
-        expected.Compliance = Compliance(
-            Status="PASS" + "ED",
-            RelatedRequirements=[],
-            AssociatedStandards=[],
-        )
-
-        # Set the check's remediation
-        expected.Remediation = {
-            "Recommendation": finding.check_metadata.Remediation.Recommendation,
-            # "Code": finding.check_metadata.Remediation.Code,
-        }
-
-        expected.Remediation["Recommendation"].Text = (
-            finding.check_metadata.Remediation.Recommendation.Text
-        )
-        expected.Remediation["Recommendation"].Url = (
-            "https://docs.aws.amazon.com/securityhub/latest/userguide/what-is-securityhub.html"
-        )
-
-        input = Check_Output_JSON_ASFF()
-        output_options = mock.MagicMock()
-
-        assert (
-            fill_json_asff(input, input_audit_info, finding, output_options) == expected
-        )
-
-    def test_fill_json_asff_with_long_associated_standards(self):
-        input_audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
-        with patch(
-            "prowler.lib.outputs.json.get_check_compliance",
-            return_value={
-                "CISA": ["your-systems-3", "your-data-2"],
-                "SOC2": ["cc_2_1", "cc_7_2", "cc_a_1_2"],
-                "CIS-1.4": ["3.1"],
-                "CIS-1.5": ["3.1"],
-                "GDPR": ["article_25", "article_30"],
-                "AWS-Foundational-Security-Best-Practices": ["cloudtrail"],
-                "HIPAA": [
-                    "164_308_a_1_ii_d",
-                    "164_308_a_3_ii_a",
-                    "164_308_a_6_ii",
-                    "164_312_b",
-                    "164_312_e_2_i",
-                ],
-                "ISO27001": ["A.12.4"],
-                "GxP-21-CFR-Part-11": ["11.10-e", "11.10-k", "11.300-d"],
-                "AWS-Well-Architected-Framework-Security-Pillar": [
-                    "SEC04-BP02",
-                    "SEC04-BP03",
-                ],
-                "GxP-EU-Annex-11": [
-                    "1-risk-management",
-                    "4.2-validation-documentation-change-control",
-                ],
-                "NIST-800-171-Revision-2": [
-                    "3_1_12",
-                    "3_3_1",
-                    "3_3_2",
-                    "3_3_3",
-                    "3_4_1",
-                    "3_6_1",
-                    "3_6_2",
-                    "3_13_1",
-                    "3_13_2",
-                    "3_14_6",
-                    "3_14_7",
-                ],
-                "NIST-800-53-Revision-4": [
-                    "ac_2_4",
-                    "ac_2",
-                    "au_2",
-                    "au_3",
-                    "au_12",
-                    "cm_2",
-                ],
-                "NIST-800-53-Revision-5": [
-                    "ac_2_4",
-                    "ac_3_1",
-                    "ac_3_10",
-                    "ac_4_26",
-                    "ac_6_9",
-                    "au_2_b",
-                    "au_3_1",
-                    "au_3_a",
-                    "au_3_b",
-                    "au_3_c",
-                    "au_3_d",
-                    "au_3_e",
-                    "au_3_f",
-                    "au_6_3",
-                    "au_6_4",
-                    "au_6_6",
-                    "au_6_9",
-                    "au_8_b",
-                    "au_10",
-                    "au_12_a",
-                    "au_12_c",
-                    "au_12_1",
-                    "au_12_2",
-                    "au_12_3",
-                    "au_12_4",
-                    "au_14_a",
-                    "au_14_b",
-                    "au_14_3",
-                    "ca_7_b",
-                    "cm_5_1_b",
-                    "cm_6_a",
-                    "cm_9_b",
-                    "ia_3_3_b",
-                    "ma_4_1_a",
-                    "pm_14_a_1",
-                    "pm_14_b",
-                    "pm_31",
-                    "sc_7_9_b",
-                    "si_1_1_c",
-                    "si_3_8_b",
-                    "si_4_2",
-                    "si_4_17",
-                    "si_4_20",
-                    "si_7_8",
-                    "si_10_1_c",
-                ],
-                "ENS-RD2022": [
-                    "op.acc.6.r5.aws.iam.1",
-                    "op.exp.5.aws.ct.1",
-                    "op.exp.8.aws.ct.1",
-                    "op.exp.8.aws.ct.6",
-                    "op.exp.9.aws.ct.1",
-                    "op.mon.1.aws.ct.1",
-                ],
-                "NIST-CSF-1.1": [
-                    "ae_1",
-                    "ae_3",
-                    "ae_4",
-                    "cm_1",
-                    "cm_3",
-                    "cm_6",
-                    "cm_7",
-                    "am_3",
-                    "ac_6",
-                    "ds_5",
-                    "ma_2",
-                    "pt_1",
-                ],
-                "RBI-Cyber-Security-Framework": ["annex_i_7_4"],
-                "FFIEC": [
-                    "d2-ma-ma-b-1",
-                    "d2-ma-ma-b-2",
-                    "d3-dc-an-b-3",
-                    "d3-dc-an-b-4",
-                    "d3-dc-an-b-5",
-                    "d3-dc-ev-b-1",
-                    "d3-dc-ev-b-3",
-                    "d3-pc-im-b-3",
-                    "d3-pc-im-b-7",
-                    "d5-dr-de-b-3",
-                ],
-                "PCI-3.2.1": ["cloudtrail"],
-                "FedRamp-Moderate-Revision-4": [
-                    "ac-2-4",
-                    "ac-2-g",
-                    "au-2-a-d",
-                    "au-3",
-                    "au-6-1-3",
-                    "au-12-a-c",
-                    "ca-7-a-b",
-                    "si-4-16",
-                    "si-4-2",
-                    "si-4-4",
-                    "si-4-5",
-                ],
-                "FedRAMP-Low-Revision-4": ["ac-2", "au-2", "ca-7"],
-            },
-        ):
-            finding = Check_Report(
-                load_check_metadata(
-                    f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
-                ).json()
-            )
-
-            # Empty the Remediation.Recomendation.URL
-            finding.check_metadata.Remediation.Recommendation.Url = ""
-
-            finding.resource_details = "Test resource details"
-            finding.resource_id = "test-resource"
-            finding.resource_arn = "test-arn"
-            finding.region = "eu-west-1"
-            finding.status = "PASS"
-            finding.status_extended = "This is a test"
-
-            expected = Check_Output_JSON_ASFF()
-            expected.Id = f"prowler-{finding.check_metadata.CheckID}-123456789012-eu-west-1-{hash_sha512('test-resource')}"
-            expected.ProductArn = (
-                "arn:aws:securityhub:eu-west-1::product/prowler/prowler"
-            )
-            expected.ProductFields = ProductFields(
-                ProviderVersion=prowler_version, ProwlerResourceName="test-arn"
-            )
-            expected.GeneratorId = "prowler-" + finding.check_metadata.CheckID
-            expected.AwsAccountId = AWS_ACCOUNT_ID
-            expected.Types = finding.check_metadata.CheckType
-            expected.FirstObservedAt = expected.UpdatedAt = expected.CreatedAt = (
-                timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
-            )
-            expected.Severity = Severity(Label=finding.check_metadata.Severity.upper())
-            expected.Title = finding.check_metadata.CheckTitle
-            expected.Description = finding.status_extended
-            expected.Resources = [
-                Resource(
-                    Id="test-arn",
-                    Type=finding.check_metadata.ResourceType,
-                    Partition="aws",
-                    Region="eu-west-1",
-                )
-            ]
-
-            expected.Compliance = Compliance(
-                Status="PASS" + "ED",
-                RelatedRequirements=[
-                    "CISA your-systems-3 your-data-2",
-                    "SOC2 cc_2_1 cc_7_2 cc_a_1_2",
-                    "CIS-1.4 3.1",
-                    "CIS-1.5 3.1",
-                    "GDPR article_25 article_30",
-                    "AWS-Foundational-Security-Best-Practices cloudtrail",
-                    "HIPAA 164_308_a_1_ii_d 164_308_a_3_ii_a 164_308_a_6_ii 164_312_",
-                    "ISO27001 A.12.4",
-                    "GxP-21-CFR-Part-11 11.10-e 11.10-k 11.300-d",
-                    "AWS-Well-Architected-Framework-Security-Pillar SEC04-BP02 SEC04",
-                    "GxP-EU-Annex-11 1-risk-management 4.2-validation-documentation-",
-                    "NIST-800-171-Revision-2 3_1_12 3_3_1 3_3_2 3_3_3 3_4_1 3_6_1 3_",
-                    "NIST-800-53-Revision-4 ac_2_4 ac_2 au_2 au_3 au_12 cm_2",
-                    "NIST-800-53-Revision-5 ac_2_4 ac_3_1 ac_3_10 ac_4_26 ac_6_9 au_",
-                    "ENS-RD2022 op.acc.6.r5.aws.iam.1 op.exp.5.aws.ct.1 op.exp.8.aws",
-                    "NIST-CSF-1.1 ae_1 ae_3 ae_4 cm_1 cm_3 cm_6 cm_7 am_3 ac_6 ds_5 ",
-                    "RBI-Cyber-Security-Framework annex_i_7_4",
-                    "FFIEC d2-ma-ma-b-1 d2-ma-ma-b-2 d3-dc-an-b-3 d3-dc-an-b-4 d3-dc",
-                    "PCI-3.2.1 cloudtrail",
-                    "FedRamp-Moderate-Revision-4 ac-2-4 ac-2-g au-2-a-d au-3 au-6-1-",
-                ],
-                AssociatedStandards=[
-                    {"StandardsId": "CISA"},
-                    {"StandardsId": "SOC2"},
-                    {"StandardsId": "CIS-1.4"},
-                    {"StandardsId": "CIS-1.5"},
-                    {"StandardsId": "GDPR"},
-                    {"StandardsId": "AWS-Foundational-Security-Best-Practices"},
-                    {"StandardsId": "HIPAA"},
-                    {"StandardsId": "ISO27001"},
-                    {"StandardsId": "GxP-21-CFR-Part-11"},
-                    {"StandardsId": "AWS-Well-Architected-Framework-Security-Pillar"},
-                    {"StandardsId": "GxP-EU-Annex-11"},
-                    {"StandardsId": "NIST-800-171-Revision-2"},
-                    {"StandardsId": "NIST-800-53-Revision-4"},
-                    {"StandardsId": "NIST-800-53-Revision-5"},
-                    {"StandardsId": "ENS-RD2022"},
-                    {"StandardsId": "NIST-CSF-1.1"},
-                    {"StandardsId": "RBI-Cyber-Security-Framework"},
-                    {"StandardsId": "FFIEC"},
-                    {"StandardsId": "PCI-3.2.1"},
-                    {"StandardsId": "FedRamp-Moderate-Revision-4"},
-                ],
-            )
-
-            # Set the check's remediation
-            expected.Remediation = {
-                "Recommendation": finding.check_metadata.Remediation.Recommendation,
-                # "Code": finding.check_metadata.Remediation.Code,
-            }
-
-            expected.Remediation["Recommendation"].Text = (
-                finding.check_metadata.Remediation.Recommendation.Text
-            )
-            expected.Remediation["Recommendation"].Url = (
-                "https://docs.aws.amazon.com/securityhub/latest/userguide/what-is-securityhub.html"
-            )
-
-            input = Check_Output_JSON_ASFF()
-            output_options = mock.MagicMock()
-
-            assert (
-                fill_json_asff(input, input_audit_info, finding, output_options)
-                == expected
-            )
-
-    def test_fill_json_ocsf_iso_format_timestamp(self):
-        input_audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
-        finding = Check_Report(
-            load_check_metadata(
-                f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
-            ).json()
-        )
-        finding.resource_details = "Test resource details"
-        finding.resource_id = "test-resource"
-        finding.resource_arn = "test-arn"
-        finding.region = "eu-west-1"
-        finding.status = "PASS"
-        finding.status_extended = "This is a test"
-
-        expected = Check_Output_JSON_OCSF(
-            finding=Finding(
-                title="Ensure Access Keys unused are disabled",
-                desc="Ensure Access Keys unused are disabled",
-                supporting_data={
-                    "Risk": "Risk associated.",
-                    "Notes": "additional information",
-                },
-                remediation=Remediation_OCSF(
-                    kb_articles=[
-                        "code or URL to the code location.",
-                        "code or URL to the code location.",
-                        "cli command or URL to the cli command location.",
-                        "cli command or URL to the cli command location.",
-                        "https://myfp.com/recommendations/dangerous_things_and_how_to_fix_them.html",
-                    ],
-                    desc="Run sudo yum update and cross your fingers and toes.",
-                ),
-                types=["Software and Configuration Checks"],
-                src_url="https://serviceofficialsiteorpageforthissubject",
-                uid="prowler-aws-iam_user_accesskey_unused-123456789012-eu-west-1-test-resource",
-                related_events=[
-                    "othercheck1",
-                    "othercheck2",
-                    "othercheck3",
-                    "othercheck4",
-                ],
-            ),
-            resources=[
-                Resources(
-                    group=Group(name="iam"),
-                    region="eu-west-1",
-                    name="test-resource",
-                    uid="test-arn",
-                    labels=[],
-                    type="AwsIamAccessAnalyzer",
-                    details="Test resource details",
-                )
-            ],
-            status_detail="This is a test",
-            compliance=Compliance_OCSF(
-                status="Success", requirements=[], status_detail="This is a test"
-            ),
-            message="This is a test",
-            severity_id=2,
-            severity="Low",
-            cloud=Cloud(
-                account=Account(name="", uid="123456789012"),
-                region="eu-west-1",
-                org=Organization(uid="", name=""),
-                provider="aws",
-                project_uid="",
-            ),
-            time=timestamp.isoformat(),
-            metadata=Metadata(
-                original_time=timestamp.isoformat(),
-                profiles=["default"],
-                product=Product(
-                    language="en",
-                    name="Prowler",
-                    version=prowler_version,
-                    vendor_name="Prowler/ProwlerPro",
-                    feature=Feature(
-                        name="iam_user_accesskey_unused",
-                        uid="iam_user_accesskey_unused",
-                        version=prowler_version,
-                    ),
-                ),
-                version="1.0.0-rc.3",
-            ),
-            state_id=0,
-            state="New",
-            status_id=1,
-            status="Success",
-            type_uid=200101,
-            type_name="Security Finding: Create",
-            impact_id=0,
-            impact="Unknown",
-            confidence_id=0,
-            confidence="Unknown",
-            activity_id=1,
-            activity_name="Create",
-            category_uid=2,
-            category_name="Findings",
-            class_uid=2001,
-            class_name="Security Finding",
-        )
-        output_options = mock.MagicMock()
-        output_options.unix_timestamp = False
-        assert fill_json_ocsf(input_audit_info, finding, output_options) == expected
-
-    def test_fill_json_ocsf_unix_timestamp(self):
-        input_audit_info = AWS_Audit_Info(
-            session_config=None,
-            original_session=None,
-            audit_session=None,
-            audited_account=AWS_ACCOUNT_ID,
-            audited_account_arn=f"arn:aws:iam::{AWS_ACCOUNT_ID}:root",
-            audited_identity_arn="test-arn",
-            audited_user_id="test",
-            audited_partition="aws",
-            profile="default",
-            profile_region="eu-west-1",
-            credentials=None,
-            assumed_role_info=None,
-            audited_regions=["eu-west-2", "eu-west-1"],
-            organizations_metadata=None,
-            audit_resources=None,
-            mfa_enabled=False,
-            audit_metadata=Audit_Metadata(
-                services_scanned=0,
-                expected_checks=[],
-                completed_checks=0,
-                audit_progress=0,
-            ),
-        )
-        finding = Check_Report(
-            load_check_metadata(
-                f"{path.dirname(path.realpath(__file__))}/fixtures/metadata.json"
-            ).json()
-        )
-        finding.resource_details = "Test resource details"
-        finding.resource_id = "test-resource"
-        finding.resource_arn = "test-arn"
-        finding.region = "eu-west-1"
-        finding.status = "PASS"
-        finding.status_extended = "This is a test"
-
-        expected = Check_Output_JSON_OCSF(
-            finding=Finding(
-                title="Ensure Access Keys unused are disabled",
-                desc="Ensure Access Keys unused are disabled",
-                supporting_data={
-                    "Risk": "Risk associated.",
-                    "Notes": "additional information",
-                },
-                remediation=Remediation_OCSF(
-                    kb_articles=[
-                        "code or URL to the code location.",
-                        "code or URL to the code location.",
-                        "cli command or URL to the cli command location.",
-                        "cli command or URL to the cli command location.",
-                        "https://myfp.com/recommendations/dangerous_things_and_how_to_fix_them.html",
-                    ],
-                    desc="Run sudo yum update and cross your fingers and toes.",
-                ),
-                types=["Software and Configuration Checks"],
-                src_url="https://serviceofficialsiteorpageforthissubject",
-                uid="prowler-aws-iam_user_accesskey_unused-123456789012-eu-west-1-test-resource",
-                related_events=[
-                    "othercheck1",
-                    "othercheck2",
-                    "othercheck3",
-                    "othercheck4",
-                ],
-            ),
-            resources=[
-                Resources(
-                    group=Group(name="iam"),
-                    region="eu-west-1",
-                    name="test-resource",
-                    uid="test-arn",
-                    labels=[],
-                    type="AwsIamAccessAnalyzer",
-                    details="Test resource details",
-                )
-            ],
-            status_detail="This is a test",
-            compliance=Compliance_OCSF(
-                status="Success", requirements=[], status_detail="This is a test"
-            ),
-            message="This is a test",
-            severity_id=2,
-            severity="Low",
-            cloud=Cloud(
-                account=Account(name="", uid="123456789012"),
-                region="eu-west-1",
-                org=Organization(uid="", name=""),
-                provider="aws",
-                project_uid="",
-            ),
-            time=int(mktime(timestamp.timetuple())),
-            metadata=Metadata(
-                original_time=int(mktime(timestamp.timetuple())),
-                profiles=["default"],
-                product=Product(
-                    language="en",
-                    name="Prowler",
-                    version=prowler_version,
-                    vendor_name="Prowler/ProwlerPro",
-                    feature=Feature(
-                        name="iam_user_accesskey_unused",
-                        uid="iam_user_accesskey_unused",
-                        version=prowler_version,
-                    ),
-                ),
-                version="1.0.0-rc.3",
-            ),
-            state_id=0,
-            state="New",
-            status_id=1,
-            status="Success",
-            type_uid=200101,
-            type_name="Security Finding: Create",
-            impact_id=0,
-            impact="Unknown",
-            confidence_id=0,
-            confidence="Unknown",
-            activity_id=1,
-            activity_name="Create",
-            category_uid=2,
-            category_name="Findings",
-            class_uid=2001,
-            class_name="Security Finding",
-        )
-        output_options = mock.MagicMock()
-        output_options.unix_timestamp = True
-        assert fill_json_ocsf(input_audit_info, finding, output_options) == expected
+    #     expected = Check_Output_JSON_OCSF(
+    #         finding=Finding(
+    #             title="Ensure Access Keys unused are disabled",
+    #             desc="Ensure Access Keys unused are disabled",
+    #             supporting_data={
+    #                 "Risk": "Risk associated.",
+    #                 "Notes": "additional information",
+    #             },
+    #             remediation=Remediation_OCSF(
+    #                 kb_articles=[
+    #                     "code or URL to the code location.",
+    #                     "code or URL to the code location.",
+    #                     "cli command or URL to the cli command location.",
+    #                     "cli command or URL to the cli command location.",
+    #                     "https://myfp.com/recommendations/dangerous_things_and_how_to_fix_them.html",
+    #                 ],
+    #                 desc="Run sudo yum update and cross your fingers and toes.",
+    #             ),
+    #             types=["Software and Configuration Checks"],
+    #             src_url="https://serviceofficialsiteorpageforthissubject",
+    #             uid="prowler-aws-iam_user_accesskey_unused-123456789012-eu-west-1-test-resource",
+    #             related_events=[
+    #                 "othercheck1",
+    #                 "othercheck2",
+    #                 "othercheck3",
+    #                 "othercheck4",
+    #             ],
+    #         ),
+    #         resources=[
+    #             Resources(
+    #                 group=Group(name="iam"),
+    #                 region="eu-west-1",
+    #                 name="test-resource",
+    #                 uid="test-arn",
+    #                 labels=[],
+    #                 type="AwsIamAccessAnalyzer",
+    #                 details="Test resource details",
+    #             )
+    #         ],
+    #         status_detail="This is a test",
+    #         compliance=Compliance_OCSF(
+    #             status="Success", requirements=[], status_detail="This is a test"
+    #         ),
+    #         message="This is a test",
+    #         severity_id=2,
+    #         severity="Low",
+    #         cloud=Cloud(
+    #             account=Account(name="", uid="123456789012"),
+    #             region="eu-west-1",
+    #             org=Organization(uid="", name=""),
+    #             provider="aws",
+    #             project_uid="",
+    #         ),
+    #         time=int(mktime(timestamp.timetuple())),
+    #         metadata=Metadata(
+    #             original_time=int(mktime(timestamp.timetuple())),
+    #             profiles=["default"],
+    #             product=Product(
+    #                 language="en",
+    #                 name="Prowler",
+    #                 version=prowler_version,
+    #                 vendor_name="Prowler/ProwlerPro",
+    #                 feature=Feature(
+    #                     name="iam_user_accesskey_unused",
+    #                     uid="iam_user_accesskey_unused",
+    #                     version=prowler_version,
+    #                 ),
+    #             ),
+    #             version="1.0.0-rc.3",
+    #         ),
+    #         state_id=0,
+    #         state="New",
+    #         status_id=1,
+    #         status="Success",
+    #         type_uid=200101,
+    #         type_name="Security Finding: Create",
+    #         impact_id=0,
+    #         impact="Unknown",
+    #         confidence_id=0,
+    #         confidence="Unknown",
+    #         activity_id=1,
+    #         activity_name="Create",
+    #         category_uid=2,
+    #         category_name="Findings",
+    #         class_uid=2001,
+    #         class_name="Security Finding",
+    #     )
+    #     output_options = mock.MagicMock()
+    #     output_options.unix_timestamp = True
+    #     assert fill_json_ocsf(input_audit_info, finding, output_options) == expected
 
     def test_extract_findings_statistics_different_resources(self):
         finding_1 = mock.MagicMock()
@@ -1285,20 +623,3 @@ class Test_Outputs:
             "CIS-1.4": ["2.1.3"],
             "CIS-1.5": ["2.1.3"],
         }
-
-    def test_generate_json_asff_status(self):
-        assert generate_json_asff_status("PASS") == "PASSED"
-        assert generate_json_asff_status("FAIL") == "FAILED"
-        assert generate_json_asff_status("MUTED") == "MUTED"
-        assert generate_json_asff_status("SOMETHING ELSE") == "NOT_AVAILABLE"
-
-    def test_generate_json_asff_resource_tags(self):
-        assert generate_json_asff_resource_tags(None) is None
-        assert generate_json_asff_resource_tags([]) is None
-        assert generate_json_asff_resource_tags([{}]) is None
-        assert generate_json_asff_resource_tags([{"key1": "value1"}]) == {
-            "key1": "value1"
-        }
-        assert generate_json_asff_resource_tags(
-            [{"Key": "key1", "Value": "value1"}]
-        ) == {"key1": "value1"}
